requirements.txt
absl-py==0.9.0
boto3==1.12.39
botocore==1.15.39
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
docutils==0.15.2
future==0.18.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
idna==2.9
imageio==2.8.0
jmespath==0.9.5
joblib==0.14.1
Markdown==3.2.1
numpy==1.18.2
oauthlib==3.1.0
pandas==0.25.1
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytorch-lightning==0.6.0
pytorch-nlp==0.5.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.2.post1
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
tensorboard==2.0.2
test-tube==0.7.5
torch==1.3.1
torchvision==0.4.2
tqdm==4.45.0
transformers==2.3.0
urllib3==1.25.8
Werkzeug==1.0.1